[Core] Update more of the source to work with KP's game system.
Cleaned out the logging system the project had. Was very useless on how it was implemented.

More fixes of the source so it compiles and works properly, instead of not work at all.

Fixed the command system so it works in-game and on console.

Clearing out old SQL files, cause we will apply fresh updates to the database to correct a ton of things, so the content is more correct then guessed work.